Chegg And Courshero/requirements.txt
discord.py
discord
requests
aiohttp-retry
PyPDF2
bs4
httpx
python-docx